edit cabin and abstracting react query

+++ src/services/apiCabins.js
@@ -11,50 +11,49 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-    '/',
-    ''
-  );
-
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
-
-  if (error) {
-    console.error(error);
-    throw new Error('Cabin could not be created!');
+export async function createEditCabin(newCabin, id) {
+  if (!newCabin.image) {
+    throw new Error('Image is required for the cabin.');
   }
 
-  // 2. Upload image
+  const hasImagePath =
+    typeof newCabin.image === 'string' &&
+    newCabin.image.startsWith(supabaseUrl);
+  const imageName = `${Date.now()}-${newCabin.image.name}`.replaceAll('/', '');
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  let query = supabase.from('cabins');
+  const cabinData = { ...newCabin, image: imagePath };
+
+  if (!id) {
+    query = query.insert([cabinData]);
+  } else {
+    query = query.update(cabinData).eq('id', id);
+  }
+
+  const { data, error } = await query.select().single();
+  if (error) {
+    console.error(error);
+    throw new Error('Cabin could not be created or updated!');
+  }
+
   const { error: storageError } = await supabase.storage
-    .from('cabing-images')
+    .from('cabin-images')
     .upload(imageName, newCabin.image);
 
-  // 3. Delete the cabin if there was an error uploading image
-
   if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id);
+    if (data) {
+      await supabase.from('cabins').delete().eq('id', data.id);
+    }
     console.error(storageError);
     throw new Error(
-      'Cabin image could not be Uploaded and the cabin was not created!'
+      'Cabin image upload failed, and the cabin has been deleted.'
     );
   }
 
   return data;
 }
 
-export async function deleteCabin(id) {
-  const { data, error } = await supabase.from('cabins').delete().eq('id', id);
-
-  if (error) {
-    console.error(error);
-    throw new Error('Cabin could not be deleted');
-  }
-
-  return data;
-}
+
